feat: Add TS, Eslint, Prettier and Husky to follow Google style guidelines
