+++ .prettierignore
@@ -1,0 +1,7 @@
+.next
+next-env.d.ts
+next.config.js
+node_modules
+yarn.lock
+package-lock.json
+public
